// src/app/features/payments/types.ts
export type PaymentStatus = 
  | 'pending'                 // Pendiente de pago
  | 'completed'               // Pago completado
  | 'failed'                  // Pago fallido
  | 'refunded'                // Pago reembolsado
  | 'under_review';